+++ src/main/kotlin/net/evilblock/prisonaio/module/battlepass/challenge/impl/PrestigeChallenge.kt
@@ -77,8 +77,8 @@ class PrestigeChallenge(id: String, internal var prestige: Int) : Challenge(id) 
 
         override fun startSetupPrompt(player: Player, id: String, lambda: (Challenge) -> Unit) {
             ConversationUtil.startConversation(player, NumberPrompt { number ->
-                assert(number > 0)
-                lambda.invoke(PrestigeChallenge(id, number))
+                assert(number.toInt() > 0)
+                lambda.invoke(PrestigeChallenge(id, number.toInt()))
             })
         }
     }
